% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleclass_format.R
\name{sampleclass_format}
\alias{sampleclass_format}
\title{Create sampleclass format for GeoDE/Characteristic Direction Input}
\usage{
sampleclass_format(gct_list)
}
\arguments{
\item{gct_list}{List of GCT objects returned from seek_data() function}
}
\value{
Factor assigning level 1 to control samples and 2 for perturbed samples to then be used with seek_DE()
}
\description{
Output from seek_data is used to assign control vs perturbed factor levels for datain_format for GeoDE/Characteristic Direction differential expression analysis used in seek_DE()
}
