% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seek_data.R
\name{seek_data}
\alias{seek_data}
\title{Parse LINCS level 3 LoF or OE Data}
\usage{
seek_data(ids, gctx)
}
\arguments{
\item{ids}{output from seek_ids(), list of sample ID's for parsing gctx}

\item{gctx}{filepath to LINCS level 3 gctx (from GEO, GSE92742_Broad_LINCS_Level3_INF_mlr12k_n1319138x12328.gctx)}
}
\value{
perturbation and control data as a list of two GCT objects
}
\description{
List of sample ID's from seek_ids() output is used to parse LINCS level 3 data for transcriptomic data and returns perturbation and control data as a list of two GCT objects
}
